refactor: allow for multiple graphics drivers in the future
