+++ packages/pros-graphics/Cargo.toml
@@ -12,9 +12,13 @@ categories = [
 ]
 
 [dependencies]
-embedded-graphics-core = "0.4.0"
+embedded-graphics-core = { version = "0.4.0", optional = true }
 pros-devices = { version = "0.1.0", path = "../pros-devices" }
 pros-sys = { version = "0.7.0", path = "../pros-sys" }
 
 [lints]
 workspace = true
+
+[features]
+default = []
+embedded-graphics = ["dep:embedded-graphics-core"]
